Add a GPU stats dialog
This is basically inspired by what you might see in e.g. GPU-Z or other similar tools.

Lots of refactoring to go along with this since it turns out that NVML calls won't return the actual current clock frequencies for a device, so NVAPI needed to be pulled into the project as well.