New Nvidia driver download implementation
Switched from WebClient to HttpClient since the former is considered outdated/deprecated. The code also streams the response (e.g. the driver file) to disk instead of buffering in memory first, which keeps memory usage lower and eliminates the added disk write delay that took place in the old code.