.NET 8 migration
Switching to .NET 8 provides many benefits and eliminates the need for some weird workarounds that were previously necessary. It also allows for single-exe publishing which is nice.

Also starting work on an RPC server to support changing power plans and some other functions via IPC. This would be easier for some automation to hook into than the current keyboard shortcuts. Can be enabled or disabled in config.

As a first addition, added display resolution switching. Attempting to work on an HDR switch functionality as well (still in progress).